ManhProject/requirements.txt
aws-cdk-lib==2.215.0
constructs>=10.0.0,<11.0.0
urllib3==1.26.18
boto3==1.34.0
beautifulsoup4==4.12.2
psutil==5.9.6
requests==2.31.0